requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
tensorflow-cpu==2.16.1
pandas==2.3.0
numpy==2.1.3
joblib==1.5.1
scikit-learn==1.7.0
requests==2.32.4
mlflow-skinny==3.1.1
google-auth==2.40.3
prometheus_client==0.22.1  # optional for metrics (you can remove if not needed)
gunicorn==23.0.0           # optional if running behind Gunicorn (Cloud Run can use uvicorn only)
